requirements.txt
Django~=3.1.4
djangorestframework~=3.12.2
requests~=2.25.1